main: define operators accessing parser specific fields
The operators (setter and/or getter) for each field.
The name of operators has following form:

	setter => <LANG>.<FIELD>:
	getter => :<LANG>.<FIELD>

The operators for a field are defined only if the field has setter
and getter in C language level; the operators require getValueObject
member and setValueObject member in fieldDefinition are set.
Currently, there is no parser specific field setting their members.
So, we cannot test the new code added with this commit.

In the later commit, I will implement a getter and a setter for fields
defined with --_fielddef option.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>